feat: a Home Service Robot V0
A full home service robot capable of navigating to pick up and deliver virtual objects to a destination (drop-off) location. Where add_markers node subscribes to odometry to keep track of the robot pose.

In this version goals should match (be equal) in both add_markers and pick_objects nodes.